Hela grejen med tullfria bilar som en enum? Por que?
