--- TollCalculator/src/main/java/se/kvrgic/Car.java
@@ -1,8 +1,0 @@
-package se.kvrgic;
-
-public class Car implements Vehicle {
-  @Override
-  public String getType() {
-    return "Car";
-  }
-}
