--- TollCalculator/src/main/java/se/kvrgic/Motorbike.java
@@ -1,8 +1,0 @@
-package se.kvrgic;
-
-public class Motorbike implements Vehicle {
-  @Override
-  public String getType() {
-    return "Motorbike";
-  }
-}
