+++ TollCalculator/src/main/java/se/kvrgic/TollCalculator.java
@@ -5,6 +5,8 @@ import java.util.concurrent.*;
 
 public class TollCalculator {
 
+    private List<Vehicle> tolledVehicles = Arrays.asList(Vehicle.CAR);
+    
   /**
    * Calculate the total toll fee for one day
    *
@@ -37,13 +39,7 @@ public class TollCalculator {
 
     private boolean isTollFreeVehicle(Vehicle vehicle) {
         if(vehicle == null) return false;
-        String vehicleType = vehicle.getType();
-        return vehicleType.equals(TollFreeVehicles.MOTORBIKE.getType()) ||
-               vehicleType.equals(TollFreeVehicles.TRACTOR.getType()) ||
-               vehicleType.equals(TollFreeVehicles.EMERGENCY.getType()) ||
-               vehicleType.equals(TollFreeVehicles.DIPLOMAT.getType()) ||
-               vehicleType.equals(TollFreeVehicles.FOREIGN.getType()) ||
-               vehicleType.equals(TollFreeVehicles.MILITARY.getType());
+        return !tolledVehicles.contains(vehicle);
     }
   
     public int getTollFee(final Date date, Vehicle vehicle) {
@@ -90,22 +86,5 @@ public class TollCalculator {
         return false;
     }
   
-    private enum TollFreeVehicles {
-        MOTORBIKE("Motorbike"),
-        TRACTOR("Tractor"),
-        EMERGENCY("Emergency"),
-        DIPLOMAT("Diplomat"),
-        FOREIGN("Foreign"),
-        MILITARY("Military");
-        private final String type;
-  
-        TollFreeVehicles(String type) {
-            this.type = type;
-        }
-  
-        public String getType() {
-            return type;
-        }
-    }
 }
