+++ TollCalculator/src/main/java/se/kvrgic/Vehicle.java
@@ -1,6 +1,20 @@
 package se.kvrgic;
 
-public interface Vehicle {
+public enum Vehicle {
+    CAR("Car"),
+    MOTORBIKE("Motorbike"),
+    TRACTOR("Tractor"),
+    EMERGENCY("Emergency"),
+    DIPLOMAT("Diplomat"),
+    FOREIGN("Foreign"),
+    MILITARY("Military");
+    private final String type;
 
-  public String getType();
+    Vehicle(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
 }
